Move testObservation from JavaDataFrameSuite into JavaDatasetSuite
